Allow moves on buildconfig project id updates (#27)
Instead of forcing a replacement (archiving + recreate), temporarily
allow moves to the new parent project. This will not be a permanent
change because if the old parent project is being archived in the same
update, the buildconfig will also end up being archived by TeamCity
first, as deletions happen before updates.