Add option to vcs triggers to trigger on snapshot dependency changes
VCS triggers have a checkbox "Trigger a build on changes in snapshot
dependencies". The TeamCity Go client has been updated, and can be used
in the terraform provider.

TEST=manual

Applied a pipeline with a VCS trigger to a local TeamCity before the
provider change, verified that no changes were detected after the
provider change, and `snapshot_dependency_triggers = true` can now be
set in the trigger